Report update

+++ ch.ahoegger.docbox.server.database.dev/src/main/java/ch/ahoegger/docbox/server/database/dev/initialization/EntityTableTask.java
@@ -59,7 +59,7 @@ public class EntityTableTask implements ITableTask, IEntityTable {
     sqlService.insert(statementBuilder.toString());
   }
 
-  public void createEntityRow(ISqlService sqlService, Long entityId, BigDecimal partnerId, BigDecimal postingGroupId, BigDecimal entityType, Date entityDate, BigDecimal hours, BigDecimal amount, String desc, Date billed) {
+  public void createEntityRow(ISqlService sqlService, BigDecimal entityId, BigDecimal partnerId, BigDecimal postingGroupId, BigDecimal entityType, Date entityDate, BigDecimal hours, BigDecimal amount, String desc, Date billed) {
     StringBuilder statementBuilder = new StringBuilder();
     statementBuilder.append("INSERT INTO ").append(TABLE_NAME).append(" (");
     statementBuilder.append(SqlFramentBuilder.columns(ENTITY_NR, PARTNER_NR, POSTING_GROUP_NR, ENTITY_TYPE, ENTITY_DATE, HOURS, AMOUNT, DESCRIPTION, BILLED));
